Disability adjusted life years

Calculate Years Lived with Disability (YLDs) and Disability-Adjusted Life-Years based on disability weights from the Global Burden of Disease study. To estimate YLL we assume the average life expectancy for a person aged x years taken from the UN WPP Africa profile for 2022. To estimate the expected age in a given range we assume exponentially distributed ages with the range. YLLs are calculated by multiplying deaths by the number of years an individual was expected to live past their year of death.

Usage

dalys(
  x,
  ages_as_years,
  life_expectancy,
  mild_disability_weight = 0.006,
  moderate_disability_weight = 0.051,
  severe_disability_weight = 0.133,
  clinical_episode_length = 0.01375,
  severe_episode_length = 0.04795
)

Arguments

x

Input data.frame

ages_as_years

Convert ages to be in units of years

life_expectancy

data.frame of expected years left to live. See example in data for format

mild_disability_weight

disability weight for mild malaria. Assigned to clinical cases in those over 5 years old

moderate_disability_weight

disability weight for moderate malaria. Assigned to clinical cases in those under 5 years old

severe_disability_weight

disability weight for severe malaria. Assigned to all severe cases

clinical_episode_length

average length of an episode of clinical malaria

severe_episode_length

average length of an episode of severe malaria

Details

Disability weights sourced here

Duration of disease was sourced from Winskill et al 2017 (Supplementary Table 8)

This is an approximation of YLD estimation from the GBD study; disability due to comorbid conditions such as motor impairment and anemia are excluded.

Note: weights and lengths are currently for Plasmodium falciparum.