Adjust severe (and downstream mortality) rates as a result of treatment coverage.

Adjust severe (and downstream mortality) rates as a result of treatment coverage.

Usage

treatment_scaling(x, treatment_scaler, baseline_treatment = 0)

Arguments

x

Input data.frame

treatment_scaler

The impact of treatment coverage on progression to severe disease and death. A highly uncertain parameter, the analysis by Griffin et al (2016) sampled from a uniform diastribution (0, 1).

baseline_treatment

The proportion of uncomplicated malaria cases that are effectively treated historically.