Extract basic rates from malariasimulation model output

Summarises clinical incidence `clinical`, severe incidence `severe` and mortality `mortality` rates as well as per-capita years lived with disability `yld`, years of life lost `yll` and daly `dalys`. All rates are expressed per person per day.

Usage

get_rates(
  x,
  baseline_year = 2000,
  ages_as_years = TRUE,
  scaler = 0.215,
  treatment_scaler = 0.42,
  mild_disability_weight = 0.006,
  moderate_disability_weight = 0.051,
  severe_disability_weight = 0.133,
  clinical_episode_length = 0.01375,
  severe_episode_length = 0.04795,
  life_expectancy = life_expectancy_africa,
  infer_ft = TRUE
)

Arguments

x

Input data.frame

baseline_year

Baseline year (assumes simulation starts on the first day of the year)

ages_as_years

Convert ages to be in units of years

scaler

Scaler for severe cases to deaths. The original fitted scaler for P. falciparum is 0.215. See Griffin et al (2016).

treatment_scaler

The impact of treatment coverage on progression to severe disease and death. The probability of being hospitalised for untreated cases compared to treated cases is difficult to study and is not known, but can be estimated from available data with several assumptions. Our estimate is derived from data used in a meta-analysis on the impact of delayed treatment of uncomplicated malaria on progression to severe malaria (Mousa et al, 2020, Plos Medicine), which estimated the relative risk ratio for treatment delays of 1-2, 2-3 and >3 days compared to treatment within 1 day. Here, we assume that the risk of hospitalisation with severe malaria among untreated cases would be double the risk in this last “delay category” of >3 days (2 * RR 1.186), giving a relative risk of any severe malaria of 0.578 for treatment within 1 day compared to no treatment (i.e. a treatment scaler of 0.42 for the proportion of severe malaria cases that are prevented by treatment of uncomplicated malaria). However, in sensitivity analysis a range of 0.281-0.843 should be considered for the treatment scaler, which was derived by assuming the risk among untreated cases would be the same as the risk with a treatment delay of >3 days (lower estimate) or 3 times the risk as those with a treatment delay of >3 days (upper estimate).

mild_disability_weight

disability weight for mild malaria. Assigned to clinical cases in those over 5 years old

moderate_disability_weight

disability weight for moderate malaria. Assigned to clinical cases in those under 5 years old

severe_disability_weight

disability weight for severe malaria. Assigned to all severe cases

clinical_episode_length

average length of an episode of clinical malaria

severe_episode_length

average length of an episode of severe malaria

life_expectancy

data.frame of expected years left to live. See example in data for format

infer_ft

If ft not found in model output (usually if ft = 0), assume ft = 0. If FALSE an error will be thrown

Details

For DALYs: disability weights are from the Global Burden of Disease study. To estimate YLL we assume the average life expectancy for a person aged x years taken from the UN WPP sub-Saharan-Africa profile. To estimate the expected age in a given range we assume exponentially distributed ages with the range. Disability weights sourced: here. This is an approximation of YLD estimation from the GBD study; disability due to comorbid conditions such as motor impairment and aneamia are excluded.

Note: Default parameter values are for _Plasmodium falciparum_